Use XpmStaticallyKnownFunctionProvider for enumerating functions matching an xs:QName.
