+++ src/lang-xpm/main/uk/co/reecedunn/intellij/plugin/xpm/StaticContext.kt
@@ -17,6 +17,7 @@ package uk.co.reecedunn.intellij.plugin.xpm
 
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
+import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xpm.optree.function.XpmFunctionDeclaration
 import uk.co.reecedunn.intellij.plugin.xpm.optree.function.XpmStaticallyKnownFunctionProvider
 import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmInScopeVariableProvider
@@ -29,6 +30,12 @@ fun PsiElement.inScopeVariables(): Sequence<XpmVariableDefinition> {
 }
 
 fun PsiFile.staticallyKnownFunctions(): Sequence<XpmFunctionDeclaration> {
+    return XpmStaticallyKnownFunctionProvider.EP_NAME.extensionList.asSequence().flatMap {
+        it.getInstance().staticallyKnownFunctions(this)
+    }
+}
+
+fun XsQNameValue.staticallyKnownFunctions(): Sequence<XpmFunctionDeclaration> {
     return XpmStaticallyKnownFunctionProvider.EP_NAME.extensionList.asSequence().flatMap {
         it.getInstance().staticallyKnownFunctions(this)
     }
