+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/completion/XQueryLookupElementTest.kt
@@ -37,6 +37,7 @@ import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableDeclaratio
 import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableReference
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryModule
 import uk.co.reecedunn.intellij.plugin.xquery.optree.XQueryInScopeVariableProvider
+import uk.co.reecedunn.intellij.plugin.xquery.optree.XQueryStaticallyKnownFunctionProvider
 import uk.co.reecedunn.intellij.plugin.xquery.tests.parser.ParserTestCase
 
 // NOTE: This class is private so the JUnit 4 test runner does not run the tests contained in it.
@@ -45,6 +46,7 @@ import uk.co.reecedunn.intellij.plugin.xquery.tests.parser.ParserTestCase
 private class XQueryLookupElementTest : ParserTestCase() {
     override fun registerExtensions() {
         registerInScopeVariableProvider(XQueryInScopeVariableProvider, "INSTANCE")
+        registerStaticallyKnownFunctionProvider(XQueryStaticallyKnownFunctionProvider, "INSTANCE")
     }
 
     @Nested
