+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/parser/ParserTestCase.kt
@@ -47,6 +47,8 @@ import uk.co.reecedunn.intellij.plugin.xpm.module.loader.XpmModuleLoaderFactory
 import uk.co.reecedunn.intellij.plugin.xpm.module.loader.XpmModuleLoaderSettings
 import uk.co.reecedunn.intellij.plugin.xpm.module.path.XpmModulePathFactoryBean
 import uk.co.reecedunn.intellij.plugin.xpm.module.path.impl.XpmModuleLocationPath
+import uk.co.reecedunn.intellij.plugin.xpm.optree.function.XpmStaticallyKnownFunctionProvider
+import uk.co.reecedunn.intellij.plugin.xpm.optree.function.XpmStaticallyKnownFunctionProviderBean
 import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmInScopeVariableProvider
 import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmInScopeVariableProviderBean
 
@@ -98,6 +100,10 @@ abstract class ParserTestCase :
 
         registerExtensionPoint(XpmFunctionDecorator.EP_NAME, XpmFunctionDecoratorBean::class.java)
         registerExtensionPoint(XpmInScopeVariableProvider.EP_NAME, XpmInScopeVariableProviderBean::class.java)
+        registerExtensionPoint(
+            XpmStaticallyKnownFunctionProvider.EP_NAME,
+            XpmStaticallyKnownFunctionProviderBean::class.java
+        )
 
         registerExtensions()
     }
@@ -149,6 +155,18 @@ abstract class ParserTestCase :
         bean.fieldName = fieldName
         bean.setPluginDescriptor(DefaultPluginDescriptor(PluginId.getId("registerInScopeVariables"), classLoader))
         registerExtension(XpmInScopeVariableProvider.EP_NAME, bean)
+    }
+
+    @Suppress("UsePropertyAccessSyntax")
+    protected fun registerStaticallyKnownFunctionProvider(
+        provider: XpmStaticallyKnownFunctionProvider, fieldName: String
+    ) {
+        val classLoader = ParserTestCase::class.java.classLoader
+        val bean = XpmStaticallyKnownFunctionProviderBean()
+        bean.implementationClass = provider.javaClass.name
+        bean.fieldName = fieldName
+        bean.setPluginDescriptor(DefaultPluginDescriptor(PluginId.getId("registerFunctions"), classLoader))
+        registerExtension(XpmStaticallyKnownFunctionProvider.EP_NAME, bean)
     }
 
     protected val settings: XQueryProjectSettings
